requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.6
dj-database-url==2.1.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-rest-framework-nested==0.0.1
djangorestframework==3.14.0
drf-spectacular==0.26.5
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
rpds-py==0.10.6
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
requests==2.31.0

